Moved the search and categories filters to the primary nav
